requirements.txt
absl-py==1.0.0
altair==4.2.0
anyio==3.5.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
astor==0.8.1
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
bleach==4.1.0
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.2
fastapi==0.74.1
flatbuffers==2.0
gast==0.5.3
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
gunicorn==20.1.0
h11==0.13.0
h5py==3.6.0
httptools==0.3.0
idna==3.3
importlib-metadata==4.11.1
ipykernel==6.9.1
ipython==8.1.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.22.2
oauthlib==3.2.0
opencv-python-headless==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
pydeck==0.7.1
Pygments==2.11.2
Pympler==1.0.1
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
stack-data==0.2.0
starlette==0.17.1
streamlit==1.6.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.13.1
testpath==0.6.0
tf-estimator-nightly==2.8.0.dev2021122109
toml==0.10.2
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
uvloop==0.16.0
validators==0.18.2
watchgod==0.7
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.2
Werkzeug==2.0.3
widgetsnbextension==3.5.2
wrapt==1.13.3
zipp==3.7.0
uvicorn==0.17.5
